requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.110.0
google-auth==2.48.0
google-genai==1.65.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
Mako==1.3.10
MarkupSafe==3.0.3
openai==2.24.0
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
gunicorn==21.2.0
uvicorn[standard]==0.24.0.post1
